moved all firebase interactions to 'services' folder
